Add an embedded minimal HTTP/1.1 server
This is for unit tests of applications built on top of httpbridge.

Doing this work has made me realize that I need to alter the primary
interface somewhat of how somebody implements an httpbridge backend.

I think I need to expose a "Frame" to the user, because calling
a body chunk a "Request" is misleading and confusing.