Add a new meta HTTP status code to say "will send response later"
This is used by the IMQS HTTP dispatcher to be informed by a handler that
the handler is going to start sending the response from another thread.
This new enum is used to distinguish the above scenario, from the case
where a code path simply fails to send a response, when it actually meant
to.